Fix build, rename Stem script
This fixes the linking errors found the final steps of the deb build process
by setting the RPATH to look for onions-common and its components in the
typical installation path. This allows it to be flexible when compiling
manually while still allowing system paths during .deb assembly.

This commit also renames the Stem script and installs it into /usr/bin
rather than /var/lib.